fix(cli): add flag to work around moving the last space from display
Yabai does not allow moving the only space from the display. It needs
some space to exist in the background.

The `--create-extra-space-if-last-on-display` is a workaround that, in
this situation, creates an extra space, so it can become the active one
on that display, prior to moving the active space to another display