fix(cli): spaces out of order after creating space during focus
Problem: When creating the target space when attempting to focus it, the
space appeared as the last one in the display. It was not reordered.

Solution: Reorder the spaces using stable indexes after focusing spaces.

Possible enhancement: Only reorder spaces when a new one was actually
created. Spaces do not have to be reordered when an existing space was
focused.